Updated screen for non UCSD wifi
